:bug: Fix(solc): Fix typescript types
