+++ bundler/solc/src/types.ts
@@ -5,6 +5,37 @@ import type {
 } from './solcTypes.js'
 import type { ModuleInfo } from '@evmts/resolutions'
 import type { Node } from 'solidity-ast/node.js'
+import type { ResolvedCompilerConfig } from '@evmts/config'
+
+export type ResolveArtifacts = (
+	solFile: string,
+	basedir: string,
+	logger: Logger,
+	config: ResolvedCompilerConfig,
+	includeAst: boolean,
+	fao: FileAccessObject,
+) => Promise<{
+	artifacts: Artifacts
+	modules: Record<'string', ModuleInfo>
+	asts: Record<string, Node> | undefined
+	solcInput: SolcInputDescription
+	solcOutput: SolcOutput
+}>
+
+export type ResolveArtifactsSync = (
+	solFile: string,
+	basedir: string,
+	logger: Logger,
+	config: ResolvedCompilerConfig,
+	includeAst: boolean,
+	fao: FileAccessObject,
+) => {
+	artifacts: Artifacts
+	modules: Record<'string', ModuleInfo>
+	asts: Record<string, Node> | undefined
+	solcInput: SolcInputDescription
+	solcOutput: SolcOutput
+}
 
 export type FileAccessObject = {
 	readFile: (path: string, encoding: BufferEncoding) => Promise<string>
